Haroon dev

+++ resources/views/app/roles/actions.blade.php
@@ -3,17 +3,23 @@
         <i class="bi bi-three-dots-vertical" style="font-size: 1rem;"></i>
     </button>
     <div class="dropdown-menu dropdown-menu-end">
-        <a class="dropdown-item" href="{{ route('roles.edit', ['id' => encryptParams($id)]) }}">
-            <i class="bi bi-pencil" style="font-size: 1.1rem" class="me-50"></i>
-            <span>{{ __('lang.commons.edit') }}</span>
-        </a>
-        <a class="dropdown-item" onclick="deleteByID('{{ encryptParams($id) }}')" href="javascript:void(0);">
-            <i class="bi bi-trash" style="font-size: 1.1rem" class="me-50"></i>
-            <span>{{ __('lang.commons.delete') }}</span>
-        </a>
-        <a class="dropdown-item" href="{{ route('roles.make-default', ['id' => encryptParams($id)]) }}">
-            <i class="bi bi-check2-circle" style="font-size: 1.1rem" class="me-50"></i>
-            <span>{{ __('lang.commons.make_default') }}</span>
-        </a>
+        @can('roles.edit')
+            <a class="dropdown-item" href="{{ route('roles.edit', ['id' => encryptParams($id)]) }}">
+                <i class="bi bi-pencil" style="font-size: 1.1rem" class="me-50"></i>
+                <span>{{ __('lang.commons.edit') }}</span>
+            </a>
+        @endcan
+        @can('roles.destroy')
+            <a class="dropdown-item" onclick="deleteByID('{{ encryptParams($id) }}')" href="javascript:void(0);">
+                <i class="bi bi-trash" style="font-size: 1.1rem" class="me-50"></i>
+                <span>{{ __('lang.commons.delete') }}</span>
+            </a>
+        @endcan
+        @can('roles.make-default')
+            <a class="dropdown-item" href="{{ route('roles.make-default', ['id' => encryptParams($id)]) }}">
+                <i class="bi bi-check2-circle" style="font-size: 1.1rem" class="me-50"></i>
+                <span>{{ __('lang.commons.make_default') }}</span>
+            </a>
+        @endcan
     </div>
 </div>
